lib/au3te-ts-tsxrs/BasePushedAuthReqEndpoint.test.ts
import { describe, expect, it, vitest } from 'vitest';
// import { AuthleteApi } from 'your-authlete-api-library';
// import { Params } from 'your-params-library';
// import { Response } from 'your-response-library';
import AuthleteApi from '../au3te-ts-common/api/AuthleteApi';
import { BasePushedAuthReqEndpoint } from './BasePushedAuthReqEndpoint';
import { Params } from './PushedAuthReqHandler';

describe('BasePushedAuthReqEndpoint', () => {
  describe('handle', () => {
    it('should call the handle method of PushedAuthReqHandler and return the response', async () => {
      // Arrange
      const api: AuthleteApi = new AuthleteApi(); // Create an instance of AuthleteApi
      const params: Params = new Params(); // Create an instance of Params
      const expectedResponse: Response = new Response(); // Create an instance of Response
      const endpoint = new BasePushedAuthReqEndpoint();

      // Mock the PushedAuthReqHandler class
      class MockPushedAuthReqHandler {
        async handle(params: Params): Promise<Response> {
          // Assert that the handle method is called with the correct params
          expect(params).toBe(params);

          // Return the expected response
          return expectedResponse;
        }
      }

      // Replace the original PushedAuthReqHandler with the mock
      vitest.spyOn(endpoint, 'handle').mockImplementation(async () => {
        const mockHandler = new MockPushedAuthReqHandler();
        return await mockHandler.handle(params);
      });

      // Act
      const actualResponse = await endpoint.handle(api, params);

      // Assert
      expect(actualResponse).toBe(expectedResponse);
    });

    it('should call the onError method and return a 500 response when an error occurs', async () => {
      // Arrange
      const api: AuthleteApi = new AuthleteApi(); // Create an instance of AuthleteApi
      const params: Params = new Params(); // Create an instance of Params
      const expectedErrorMessage = 'An error occurred';
      const endpoint = new BasePushedAuthReqEndpoint();

      // Mock the PushedAuthReqHandler class to throw an error
      class MockPushedAuthReqHandler {
        async handle(params: Params): Promise<Response> {
          throw new Error(expectedErrorMessage);
        }
      }

      // Replace the original PushedAuthReqHandler with the mock
      vitest.spyOn(endpoint, 'handle').mockImplementation(async () => {
        const mockHandler = new MockPushedAuthReqHandler();
        return await mockHandler.handle(params);
      });

      // Mock the onError method
      vitest.spyOn(endpoint, 'onError' as any); // Fix the access issue

      // Act
      const actualResponse = await endpoint.handle(api, params);

      // Assert
      expect(endpoint.onError).toHaveBeenCalledWith(
        new Error(expectedErrorMessage)
      );
      expect(actualResponse.status).toBe(500);
      expect(actualResponse.body).toBe(expectedErrorMessage);
    });
  });
});
